refactor(rx51-server-organization): melhora e organiza estrutura, padrões de projetos, rotas
