+++ src/shared/infra/app.ts
@@ -1,20 +1,30 @@
+import "reflect-metadata";
+import "dotenv/config";
 import express, { NextFunction, Request, Response } from "express";
+import "express-async-errors";
 import swaggerUi from "swagger-ui-express";
 
+import "@shared/container";
+import upload from "@config/upload";
 import { AppError } from "@shared/errors/AppError";
+import createConnection from "@shared/infra/typeorm";
 
-import { router } from "./routes";
-import swaggerFile from "./swagger.json";
+import swaggerFile from "../../swagger.json";
+import { router } from "./http/routes";
 
-import "./database";
-
+createConnection();
 const app = express();
+
 
 app.use(express.json());
 
 app.use("/api-docs", swaggerUi.serve, swaggerUi.setup(swaggerFile));
 
+app.use("/avatar", express.static(`${upload.tmpFolder}/avatar`));
+app.use("/cars", express.static(`${upload.tmpFolder}/cars`));
+
 app.use(router);
+
 
 app.use(
   (err: Error, request: Request, response: Response, next: NextFunction) => {
@@ -30,4 +40,4 @@ app.use(
   }
 );
 
-app.listen(3333, () => console.log("Sever is running!"));
+export { app };
