+++ src/shared/infra/http/middlewares/ensureAuthenticated.ts
