+++ src/shared/infra/server.ts
@@ -1,0 +1,3 @@
+import { app } from "./app";
+
+app.listen(3333, () => console.log("Server is running!"));
